Work on cross validation
Smoothing is now done on the pointwise autocorrelations instead of
separately on the variance and autocovariance, which seems to reduce
the problem of cross validation selecting a too small bandwidth. That
still seems to be a problem when the autocorrelation is high. There
are now tests with low and high autocorrelations that demonstrate
that.