Remove support for cross-validation of bandwidth
In general, supporting cross-validation to select bandwidth seems more
complicated than expected. It is necessary for the test data to be
independent of the training data while at the same time being close
enough to the training data that the kernel around the test data
includes the training data. This seems to make cross-validation
impossible if the true bandwidth is not sufficiently given the level
of autocorrelation in the time series. With multiple time series, one
could easily do cross validation by using different time series for
the test and training sets. However, as no one has requested such a
feature, we will not spend time working on it now.