Add function to create pomp gillespie simulator
The new function create a pomp object that can simulate according do
Gillepies direct method an SIR model with demography. The transmission
rate can depend on time via a sinusoidal function, and the
immigration/birth rate can also vary with time according to a ramp
function. Documentation of the parameters will be forthcoming.

The code was developed using the SIR model in the pomp package as a
template, but has significant differences. Seasonality occurs in this
model due to a simple sin function as opposed to a basis of
B-splines. The ramp function for the birth rate is unique to our
version. In the future, pomp may support Csnippets for creation of
Gillespie simulators (it's on the todo list at the moment). When that
happens, it may be worth using that facility to reduce the amount of C
code a little.

I also changed the updating matrix in the call to gillespie.sim()
compared to the sir example in the pomp package. I have set the birth
rate and transmission rate to be recalculated after any
transition. The rationale for this is that they both in general may
depend on time and thus also require some updating. It may be worth
considering having the code print a warning if the update is very
large, because that would indicate that the code is not approximating
the time-dependence of the rates very well. On the other hand, maybe
we should find an algorithm that is designed for exact simulation with
time-dependent rates. Finally, we note that it may be worth checking
how much performance suffers from this change when rates are not
time-dependent. If there's a noticeable difference, we could easily
have the updating matrix be set appropriately.