Correct and revise vignette
The equations for statistics have been corrected regarding the use of
residuals vs. the original time series. The bandwidth has been
increased in the example analyis as it seems that the previous
bandwidth was overfitting the data and hence obscuring all of the
expected trends in the warning signals. The local linear model was
removed from the example because it is slow and does not produce
noticeably different results in this case.