Correct implementation of frequency-dependent transmission
In previous commit, the rate of both the transmission and the sparking
reactions were divided by the population density. Currently, only the
transmission rate is divided by population density. Also, a bug in
create_simulator() function has been corrected and a new argument
added to control which transmission model is used.