Add stat of first difference of variance estimates
This statistic has been added to the output of get_stats to support
using the convexity of the variance as an early warning signal. A
definition is included in the vignette and it has also been
substituted into the plots in the vignette in the place of
kurtosis. This change makes it necessary to use the full name
"variance" instead of "var" to pull out the variance estimates from
the stats list of the get_stats output, which has lead to some changes
in the examples and vignette.