Eliminate error caused by a constant time series
Providing a constant time series as input to get_stats led to an
error. This was because the autocorrelation and decay time estimates
were all undefined. Providing such a vector to stats::cor() leads to
an error: "no complete element pairs". Therefore, in get_tau() we now
check for such vectors and return a missing value instead of calling
stats::cor().