Change simator parameter beta to beta_par
This was necessary when using the C snippet facility to avoid a
conflict with the global function beta. The C snippets allows for a
major simplification of the code, including a complete removal of the
src directory from spaero. The C snippet approach will hopefully
require less maintenance in going forward.