Document new option to turn off TDAR calculation
This option was added for the use case that one may want to calculate
the EWS over a grid of bandwidth values. Since the TDAR will be the
same, being that it is independent of bandwidth, it makes since to
turn off its calculation for all but one of the bandwidth values for
the sake of speed.